+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add require('@openveo/api').fileSystem.readFile to read part of a file
 - Add require('@openveo/api').fileSystem.getFileTypeFromBuffer to get the type of a file as a buffer
 - Add require('@openveo/api').util.validateFiles to validate that files types are as expected
+- Add require('@openveo/api').util.getPropertyFromArray to retrieve values of a property inside an array of objects
 
 # 4.0.0 / 2017-05-04